+++ etna/clustering/hierarchical/base.py
@@ -5,6 +5,7 @@ from typing import Optional
 from typing import Union
 
 import pandas as pd
+from sklearn import __version__ as sklearn_version
 from sklearn.cluster import AgglomerativeClustering
 
 from etna.clustering.base import Clustering
@@ -81,9 +82,16 @@ class HierarchicalClustering(Clustering):
         """
         self.n_clusters = n_clusters
         self.linkage = ClusteringLinkageMode(linkage).name
+
+        sklearn_version_tuple = tuple(map(int, sklearn_version.split(".")))
+        if sklearn_version_tuple < (1, 2):
+            clustering_algo_params["affinity"] = "precomputed"
+        else:
+            clustering_algo_params["metric"] = "precomputed"
         self.clustering_algo = AgglomerativeClustering(
-            n_clusters=self.n_clusters, affinity="precomputed", linkage=self.linkage, **clustering_algo_params
+            n_clusters=self.n_clusters, linkage=self.linkage, **clustering_algo_params
         )
+
         self.clusters = None
         self.segment2cluster = None
         self.centroids_df = None